requirements.txt
fastapi>=0.100.0
pydantic==2.0.2
uvicorn==0.22.0
mysql-connector-python==8.0.33
SQLAlchemy==2.0.18
requests==2.31.0
boto3==1.26.155
python-multipart==0.0.6
ffmpeg-python==0.2.0
aiomysql>=0.0.20
